Added missing cast.

+++ Dove/Dove.swift
@@ -19,5 +19,5 @@ public let animationConstant = 0.3
 public let shorterAnimationConstant = animationConstant / 2
 public let longerAnimationConstant = animationConstant * 2
 
-public let springDamping = 0.5
-public let springInitialVelocity = 0.1
+public let springDamping = CGFloat(0.5)
+public let springInitialVelocity = CGFloat(0.1)
